Refactor debug util
