+++ prototype/bots/story_agent.py
@@ -1,8 +1,8 @@
-import os
+
 from openai import OpenAI
 from dotenv import load_dotenv
 import json
-from utils.debug_util import debug_log
+from prototype.debug.debug_util import debug_log
 
 load_dotenv()
 client = OpenAI()
