+++ prototype/services/campaign_manager.py
@@ -1,4 +1,4 @@
-import os
+
 from db.db import (create_campaign, list_campaigns, get_most_recent_campaign, 
                    update_campaign_last_played, get_or_create_user)
